refactor: update import paths for EntityCard and ProgressEscrow to use consistent directory structure
