+++ src/components/modules/escrow/ui/dialogs/EscrowDetailDialog.tsx
@@ -16,7 +16,7 @@ import { cn } from "@/lib/utils";
 import { useFormatUtils } from "@/utils/hook/format.hook";
 import TooltipInfo from "@/components/utils/Tooltip";
 import { useCopyUtils } from "@/utils/hook/copy.hook";
-import EntityCard from "./components/EntityCard";
+import EntityCard from "./cards/EntityCard";
 import FundEscrowDialog from "./FundEscrowDialog";
 import { useEscrowBoundedStore } from "../../store/ui";
 import { useGlobalBoundedStore } from "@/core/store/data";
@@ -26,7 +26,7 @@ import { Badge } from "@/components/ui/badge";
 import useDistributeEarningsEscrowDialogHook from "./hooks/distribute-earnings-escrow-dialog.hook";
 import useChangeStatusEscrowDialogHook from "./hooks/change-status-escrow-dialog.hook";
 import useChangeFlagEscrowDialogHook from "./hooks/change-flag-escrow-dialog.hook";
-import ProgressEscrow from "./components/ProgressEscrow";
+import ProgressEscrow from "./utils/ProgressEscrow";
 import useStartDisputeEscrowDialogHook from "./hooks/start-dispute-escrow-dialog.hook";
 import ResolveDisputeEscrowDialog from "./ResolveDisputeEscrowDialog";
 import useResolveDisputeEscrowDialogHook from "./hooks/resolve-dispute-escrow-dialog.hook";
