+++ src/components/modules/escrow/ui/dialogs/cards/EntityCard.tsx
